// src/app/core/models/termsAndConditionsDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface TermsAndConditionsDto {
  id: string;
  tenantId: string;
  title: string;
  content: string;
  version: string;
  isActive: boolean;
  effectiveDate: string;
  expiryDate: string;
  createdAt: string;
  updatedAt: string;
}